Unify package name and integrate extension
