--- pyrssa/__init__.py
@@ -1,6 +1,0 @@
-"""Python helpers for rssa."""
-from .hankel import hankel_mv, HAVE_FAST as _HAVE_FAST
-
-__all__ = ["hankel_mv", "_HAVE_FAST"]
-
-
