+++ setup.py
@@ -3,16 +3,16 @@ import numpy
 
 ext_modules = [
     Extension(
-        'pyrssa._fast_hankel',
-        ['pyrssa/_fast_hankel.c'],
+        'py_rssa._fast_hankel',
+        ['py_rssa/py_rssa/_fast_hankel.c'],
         include_dirs=[numpy.get_include()],
         libraries=['fftw3'],
     )
 ]
 
 setup(
-    name='pyrssa',
+    name='py_rssa',
     version='0.1.0',
-    packages=['pyrssa'],
+    packages=['py_rssa'],
     ext_modules=ext_modules,
 )
